pyproject.toml
[project]
name = "playwright_github"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "abhijit.mamarde", email = "abhijit.mamarde@gmail.com"},
]
dependencies = [
    "pytest-playwright>=0.4.3",
    "pytest-xdist>=3.5.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.1.0",
    "setuptools>=69.0.3",
    "asyncio>=3.4.3",
    "pytest-repeat>=0.9.3",
    "pytest-html>=4.1.1",
    "allure-pytest>=2.13.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"
